move: [report id -> photo filename list] logic to celery task; chore: clean up and improve naming
